app/Core/utils.py
from typing import Any, Dict, Optional
from datetime import datetime, timedelta
from jose import jwt

from Core.config import setting


def generate_password_reset_token(email: str) -> str:
    delta = timedelta(hours=3)
    now = datetime.utcnow()
    expires = now + delta
    exp = expires.timestamp()
    encoded_jwt = jwt.encode(
        {"exp": exp, "nbf": now, "sub": email}, setting.SECRET_KEY, algorithm="HS256",
    )
    return encoded_jwt


def verify_password_reset_token(token: str) -> Optional[str]:
    try:
        decoded_token = jwt.decode(token, setting.SECRET_KEY, algorithms=["HS256"])
        return decoded_token["sub"]
    except jwt.JWTError:
        return None